+++ src/table/lock.jsx
@@ -243,12 +243,12 @@ export default function lock(BaseComponent) {
 
         scroll() {
             const { scrollToCol = 0, scrollToRow = 0 } = this.props;
-            if (!scrollToCol && !scrollToRow) {
+            if ((!scrollToCol && !scrollToRow) || !this.bodyNode) {
                 return;
             }
             const colCellNode = this.getCellNode(0, scrollToCol);
             const rowCellNode = this.getCellNode(scrollToRow, 0);
-            const bodyNodeOffset = this.bodyNode.getBoundingClientRect();
+            const bodyNodeOffset = this.bodyNode.getBoundingClientRect() || {};
             if (colCellNode) {
                 const cellNodeoffset = colCellNode.getBoundingClientRect();
                 const scrollLeft = cellNodeoffset.left - bodyNodeOffset.left;